requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow==2.2.0
keras==2.13.1
opencv-python
h5py
imgaug
IPython[all]
protobuf
geopandas